Speed up GraphChecker::VisitInstruction
We were seeing slowdowns in graphs with many instructions within
the same block. We can use extra bookkeeping so that
StrictlyDominates is faster. This slowdown only affected debug
builds where the GraphChecker would go through the instruction
list over and over again.

The app in the bug goes from timeouting at 60 minutes, to compiling
in ~8 minutes.

Bug: 311599964
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: Compile the app in the bug with dex2oatd
Change-Id: I8a3801c635633d736a97cfd6b563399266a15d20